fix: make some little fixes
