+++ src/components/SelectScheduleDialog/SelectScheduleDialog.styles.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const StyledDialogContainer = styled.div`
+    padding-bottom: 18px;
+`;
